MGDOBR-270: add AbstractWebClientInvokerBuilder to reduce code duplication
