+++ processor/runtime/src/main/java/com/redhat/service/smartevents/processor/actions/webhook/WebhookActionInvokerBuilder.java
@@ -1,62 +1,34 @@
 package com.redhat.service.smartevents.processor.actions.webhook;
 
-import java.util.Objects;
 import java.util.Optional;
 
-import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.inject.Instance;
-import javax.inject.Inject;
 
-import com.redhat.service.smartevents.infra.auth.AbstractOidcClient;
-import com.redhat.service.smartevents.infra.auth.OidcClientConstants;
-import com.redhat.service.smartevents.infra.exceptions.definitions.platform.TechnicalBearerTokenNotConfiguredException;
 import com.redhat.service.smartevents.infra.exceptions.definitions.user.GatewayProviderException;
 import com.redhat.service.smartevents.infra.models.dto.ProcessorDTO;
 import com.redhat.service.smartevents.infra.models.gateways.Action;
+import com.redhat.service.smartevents.processor.actions.AbstractWebClientInvokerBuilder;
 import com.redhat.service.smartevents.processor.actions.ActionInvoker;
-import com.redhat.service.smartevents.processor.actions.ActionInvokerBuilder;
-
-import io.vertx.ext.web.client.WebClientOptions;
-import io.vertx.mutiny.core.Vertx;
-import io.vertx.mutiny.ext.web.client.WebClient;
 
 @ApplicationScoped
-public class WebhookActionInvokerBuilder implements WebhookAction,
-        ActionInvokerBuilder {
-
-    private WebClient client;
-
-    @Inject
-    Instance<AbstractOidcClient> oidcClients;
-
-    @Inject
-    Vertx vertx;
-
-    @PostConstruct
-    private void onPostConstruct() {
-        client = WebClient.create(vertx, new WebClientOptions().setLogActivity(true));
-    }
+public class WebhookActionInvokerBuilder extends AbstractWebClientInvokerBuilder implements WebhookAction {
 
     @Override
     public ActionInvoker build(ProcessorDTO processor, Action action) {
         String endpoint = Optional.ofNullable(action.getParameters().get(ENDPOINT_PARAM))
                 .orElseThrow(() -> buildNoEndpointException(processor));
-        if (action.getParameters().containsKey(USE_TECHNICAL_BEARER_TOKEN_PARAM)
-                && action.getParameters().get(USE_TECHNICAL_BEARER_TOKEN_PARAM).equals("true")) {
-            AbstractOidcClient abstractOidcClient =
-                    oidcClients.stream()
-                            .filter(x -> Objects.equals(x.getName(), OidcClientConstants.WEBHOOK_OIDC_CLIENT_NAME))
-                            .findFirst()
-                            .orElseThrow(() -> new TechnicalBearerTokenNotConfiguredException("A webhook action needed the webhook oidc client bean but it was not configured."));
-            return new WebhookActionInvoker(endpoint, client, abstractOidcClient);
-        }
-        return new WebhookActionInvoker(endpoint, client);
+        return requiresTechnicalBearerToken(action)
+                ? new WebhookActionInvoker(endpoint, webClient, getOidcClient())
+                : new WebhookActionInvoker(endpoint, webClient);
     }
 
-    private GatewayProviderException buildNoEndpointException(ProcessorDTO processor) {
+    private static GatewayProviderException buildNoEndpointException(ProcessorDTO processor) {
         String message = String.format("There is no endpoint specified in the parameters for Action on Processor '%s' on Bridge '%s'",
                 processor.getId(), processor.getBridgeId());
         return new GatewayProviderException(message);
+    }
+
+    private static boolean requiresTechnicalBearerToken(Action action) {
+        return Boolean.parseBoolean(action.getParameters().get(USE_TECHNICAL_BEARER_TOKEN_PARAM));
     }
 }
